feat:update route.go
1.add enableCORS function in middleware.go

2.solve:  too many return values
	have (http.HandlerFunc)
	want ()
by
This error indicates a mismatch between the return values of a function and what is expected. In your code, the enableCORS function is defined to return an http.Handler, but within the return result of http.HandlerFunc, you have a return statement. This is because http.HandlerFunc is actually a function type that takes a function as a parameter and returns a function that conforms to the http.Handler interface. In your code, http.HandlerFunc has already converted the function for you, so you do not need to return any value.